Move to Vulkan header version 0.222

+++ src/vk0.10-initswapchain.cpp
@@ -183,7 +183,7 @@ int main(int argc, char **argv)
 
     VkExtent2D swapChainExtent;
     // width and height are either both -1, or both not -1.
-    if (surfCapabilities.currentExtent.width == -1)
+    if (surfCapabilities.currentExtent.width == (uint32_t) -1)
     {
         // If the surface size is undefined, the size is set to
         // the size of the images requested.
